+++ src/run/handlers/use-cache-handler.ts
@@ -5,16 +5,18 @@ import type {
   CacheEntry,
   // only supporting latest variant (https://github.com/vercel/next.js/pull/76687)
   // first released in v15.3.0-canary.13
-  CacheHandlerV2 as CacheHandler,
+  CacheHandlerV2,
 } from 'next-with-cache-handler-v2/dist/server/lib/cache-handlers/types.js'
+import type { CacheHandler as CacheHandlerWithUpdateTag } from 'next-with-cache-handler-v3/dist/server/lib/cache-handlers/types.js'
 
 import { getLogger } from './request-context.cjs'
 import {
-  getMostRecentTagRevalidationTimestamp,
-  isAnyTagStale,
+  getMostRecentTagExpirationTimestamp,
+  isAnyTagStaleOrExpired,
   markTagsAsStaleAndPurgeEdgeCache,
+  RevalidateTagDurations,
 } from './tags-handler.cjs'
-import { getTracer } from './tracer.cjs'
+import { getTracer, withActiveSpan } from './tracer.cjs'
 
 // Most of this code is copied and adapted from Next.js default 'use cache' handler implementation
 // https://github.com/vercel/next.js/blob/84fde91e03918344c5d356986914ab68a5083462/packages/next/src/server/lib/cache-handlers/default.ts
@@ -38,6 +40,8 @@ type PrivateCacheEntry = {
 type CacheHandleLRUCache = LRUCache<string, PrivateCacheEntry>
 type PendingSets = Map<string, Promise<void>>
 
+type MultiVersionCacheHandler = CacheHandlerV2 & CacheHandlerWithUpdateTag
+
 const LRU_CACHE_GLOBAL_KEY = Symbol.for('nf-use-cache-handler-lru-cache')
 const PENDING_SETS_GLOBAL_KEY = Symbol.for('nf-use-cache-handler-pending-sets')
 const cacheHandlersSymbol = Symbol.for('@next/cache-handlers')
@@ -51,8 +55,8 @@ const extendedGlobalThis = globalThis as typeof globalThis & {
 
   // Used by Next.js to provide implementation of cache handlers
   [cacheHandlersSymbol]?: {
-    RemoteCache?: CacheHandler
-    DefaultCache?: CacheHandler
+    RemoteCache?: MultiVersionCacheHandler
+    DefaultCache?: MultiVersionCacheHandler
   }
 }
 
@@ -86,12 +90,13 @@ function getPendingSets(): PendingSets {
 const tmpResolvePendingBeforeCreatingAPromise = () => {}
 
 export const NetlifyDefaultUseCacheHandler = {
-  get(cacheKey: string): ReturnType<CacheHandler['get']> {
-    return getTracer().withActiveSpan(
+  get(cacheKey: string): ReturnType<MultiVersionCacheHandler['get']> {
+    return withActiveSpan(
+      getTracer(),
       'DefaultUseCacheHandler.get',
-      async (span): ReturnType<CacheHandler['get']> => {
+      async (span): ReturnType<MultiVersionCacheHandler['get']> => {
         getLogger().withFields({ cacheKey }).debug(`[NetlifyDefaultUseCacheHandler] get`)
-        span.setAttributes({
+        span?.setAttributes({
           cacheKey,
         })
 
@@ -105,7 +110,7 @@ export const NetlifyDefaultUseCacheHandler = {
           getLogger()
             .withFields({ cacheKey, status: 'MISS' })
             .debug(`[NetlifyDefaultUseCacheHandler] get result`)
-          span.setAttributes({
+          span?.setAttributes({
             cacheStatus: 'miss',
           })
           return undefined
@@ -120,52 +125,64 @@ export const NetlifyDefaultUseCacheHandler = {
           getLogger()
             .withFields({ cacheKey, ttl, status: 'STALE' })
             .debug(`[NetlifyDefaultUseCacheHandler] get result`)
-          span.setAttributes({
+          span?.setAttributes({
             cacheStatus: 'expired, discarded',
             ttl,
           })
           return undefined
         }
 
-        if (await isAnyTagStale(entry.tags, entry.timestamp)) {
+        const { stale, expired } = await isAnyTagStaleOrExpired(entry.tags, entry.timestamp)
+
+        if (expired) {
           getLogger()
-            .withFields({ cacheKey, ttl, status: 'STALE BY TAG' })
+            .withFields({ cacheKey, ttl, status: 'EXPIRED BY TAG' })
             .debug(`[NetlifyDefaultUseCacheHandler] get result`)
 
-          span.setAttributes({
-            cacheStatus: 'stale tag, discarded',
+          span?.setAttributes({
+            cacheStatus: 'expired tag, discarded',
             ttl,
           })
           return undefined
         }
 
+        let { revalidate, value } = entry
+        if (stale) {
+          revalidate = -1
+        }
+
         // returning entry will cause stream to be consumed
         // so we need to clone it first, so in-memory cache can
         // be used again
-        const [returnStream, newSaved] = entry.value.tee()
+        const [returnStream, newSaved] = value.tee()
         entry.value = newSaved
 
         getLogger()
-          .withFields({ cacheKey, ttl, status: 'HIT' })
+          .withFields({ cacheKey, ttl, status: stale ? 'STALE' : 'HIT' })
           .debug(`[NetlifyDefaultUseCacheHandler] get result`)
-        span.setAttributes({
-          cacheStatus: 'hit',
+        span?.setAttributes({
+          cacheStatus: stale ? 'stale' : 'hit',
           ttl,
         })
 
         return {
           ...entry,
+          revalidate,
           value: returnStream,
         }
       },
     )
   },
-  set(cacheKey: string, pendingEntry: Promise<CacheEntry>): ReturnType<CacheHandler['set']> {
-    return getTracer().withActiveSpan(
+  set(
+    cacheKey: string,
+    pendingEntry: Promise<CacheEntry>,
+  ): ReturnType<MultiVersionCacheHandler['set']> {
+    return withActiveSpan(
+      getTracer(),
       'DefaultUseCacheHandler.set',
-      async (span): ReturnType<CacheHandler['set']> => {
+      async (span): ReturnType<MultiVersionCacheHandler['set']> => {
         getLogger().withFields({ cacheKey }).debug(`[NetlifyDefaultUseCacheHandler]: set`)
-        span.setAttributes({
+        span?.setAttributes({
           cacheKey,
         })
 
@@ -180,7 +197,7 @@ export const NetlifyDefaultUseCacheHandler = {
 
         const entry = await pendingEntry
 
-        span.setAttributes({
+        span?.setAttributes({
           cacheKey,
         })
 
@@ -194,7 +211,7 @@ export const NetlifyDefaultUseCacheHandler = {
             size += Buffer.from(chunk.value).byteLength
           }
 
-          span.setAttributes({
+          span?.setAttributes({
             tags: entry.tags,
             timestamp: entry.timestamp,
             revalidate: entry.revalidate,
@@ -221,20 +238,26 @@ export const NetlifyDefaultUseCacheHandler = {
     // we would need to check more tags than current request needs
     // while blocking pipeline
   },
-  getExpiration: function (...tags: string[]): ReturnType<CacheHandler['getExpiration']> {
-    return getTracer().withActiveSpan(
+  getExpiration: function (
+    // supporting both (...tags: string[]) and (tags: string[]) signatures
+    ...notNormalizedTags: string[] | string[][]
+  ): ReturnType<MultiVersionCacheHandler['getExpiration']> {
+    return withActiveSpan(
+      getTracer(),
       'DefaultUseCacheHandler.getExpiration',
-      async (span): ReturnType<CacheHandler['getExpiration']> => {
-        span.setAttributes({
+      async (span): ReturnType<MultiVersionCacheHandler['getExpiration']> => {
+        const tags = notNormalizedTags.flat()
+
+        span?.setAttributes({
           tags,
         })
 
-        const expiration = await getMostRecentTagRevalidationTimestamp(tags)
+        const expiration = await getMostRecentTagExpirationTimestamp(tags)
 
         getLogger()
           .withFields({ tags, expiration })
           .debug(`[NetlifyDefaultUseCacheHandler] getExpiration`)
-        span.setAttributes({
+        span?.setAttributes({
           expiration,
         })
 
@@ -242,12 +265,14 @@ export const NetlifyDefaultUseCacheHandler = {
       },
     )
   },
-  expireTags(...tags: string[]): ReturnType<CacheHandler['expireTags']> {
-    return getTracer().withActiveSpan(
+  // this is for CacheHandlerV2
+  expireTags(...tags: string[]): ReturnType<MultiVersionCacheHandler['expireTags']> {
+    return withActiveSpan(
+      getTracer(),
       'DefaultUseCacheHandler.expireTags',
-      async (span): ReturnType<CacheHandler['expireTags']> => {
+      async (span): ReturnType<MultiVersionCacheHandler['expireTags']> => {
         getLogger().withFields({ tags }).debug(`[NetlifyDefaultUseCacheHandler] expireTags`)
-        span.setAttributes({
+        span?.setAttributes({
           tags,
         })
 
@@ -255,7 +280,27 @@ export const NetlifyDefaultUseCacheHandler = {
       },
     )
   },
-} satisfies CacheHandler
+  // this is for CacheHandlerV3 / Next 16
+  updateTags(
+    tags: string[],
+    durations: RevalidateTagDurations,
+  ): ReturnType<MultiVersionCacheHandler['updateTags']> {
+    return withActiveSpan(
+      getTracer(),
+      'DefaultUseCacheHandler.updateTags',
+      async (span): ReturnType<MultiVersionCacheHandler['updateTags']> => {
+        getLogger()
+          .withFields({ tags, durations })
+          .debug(`[NetlifyDefaultUseCacheHandler] updateTags`)
+        span?.setAttributes({
+          tags,
+          durations: JSON.stringify(durations),
+        })
+        await markTagsAsStaleAndPurgeEdgeCache(tags, durations)
+      },
+    )
+  },
+} satisfies MultiVersionCacheHandler
 
 export function configureUseCacheHandlers() {
   extendedGlobalThis[cacheHandlersSymbol] = {