+++ tests/fixtures/wasm-src/src/app/og-node/route.js
@@ -1,4 +1,5 @@
-import { ImageResponse } from '@vercel/og'
+// see next.config for details about 'next-og-alias'
+import { ImageResponse } from 'next-og-alias'
 
 export async function GET() {
   return new ImageResponse(<div>hi</div>, {