+++ tests/utils/next-version-helpers.mjs
@@ -41,12 +41,28 @@ export function shouldHaveAppRouterNotFoundInPrerenderManifest() {
 export function shouldHaveAppRouterGlobalErrorInPrerenderManifest() {
   // https://github.com/vercel/next.js/pull/82444
 
-  // this is not used in any stable version yet
-  return isNextCanary() && nextVersionSatisfies('>=15.5.1-canary.4')
+  return (
+    (isNextCanary() && nextVersionSatisfies('>=15.5.1-canary.4')) ||
+    nextVersionSatisfies('>=16.0.0')
+  )
 }
 
 export function hasNodeMiddlewareSupport() {
   return nextVersionSatisfies(isNextCanary() ? '>=15.2.0' : '>=15.5.0')
+}
+
+export function hasDefaultTurbopackBuilds() {
+  return nextVersionSatisfies('>=15.6.0-canary.40')
+}
+
+export function shouldHaveSlashIndexTagForIndexPage() {
+  // https://github.com/vercel/next.js/pull/84586
+  return nextVersionSatisfies('>=v15.6.0-canary.50')
+}
+
+export function isExperimentalPPRHardDeprecated() {
+  // https://github.com/vercel/next.js/pull/84280
+  return nextVersionSatisfies('>=15.6.0-canary.54')
 }
 
 /**
@@ -94,7 +110,7 @@ export async function setNextVersionInFixture(
 
   const isSemverVersion = valid(resolvedVersion)
 
-  const areNextVersionConstraintsSatisfied = await Promise.all(
+  const packageJsonsNeedUpdates = await Promise.all(
     packageJsons.map(async (packageJsonPath) => {
       const packageJson = JSON.parse(await readFile(packageJsonPath, 'utf8'))
       if (packageJson.dependencies?.next) {
@@ -114,15 +130,19 @@ export async function setNextVersionInFixture(
               `${logPrefix}⏩ Skipping '${packageJson.name}' because it requires next@${versionConstraint}`,
             )
           }
-          return false
+          return { packageJsonPath, needUpdate: false }
         }
       }
-      return true
+      return { packageJsonPath, needUpdate: true }
     }),
   )
 
-  if (areNextVersionConstraintsSatisfied.some((isSatisfied) => !isSatisfied)) {
-    // at least one next version constraint is not satisfied so we skip this fixture
+  const packageJsonsToUpdate = packageJsonsNeedUpdates
+    .filter(({ needUpdate }) => needUpdate)
+    .map(({ packageJsonPath }) => packageJsonPath)
+
+  if (packageJsonsToUpdate.length === 0) {
+    // all next version constraints are not satisfied so we skip this fixture
     return false
   }
 
@@ -138,7 +158,7 @@ export async function setNextVersionInFixture(
   }
 
   await Promise.all(
-    packageJsons.map(async (packageJsonPath) => {
+    packageJsonsToUpdate.map(async (packageJsonPath) => {
       const packageJson = JSON.parse(await readFile(packageJsonPath, 'utf8'))
       if (packageJson.dependencies?.next) {
         packageJson.dependencies.next = version